update: Update application service mesh monitor
