+++ package.json
@@ -5,7 +5,7 @@
   "license": "AGPL-3.0",
   "dependencies": {
     "@babel/polyfill": "^7.12.1",
-    "@kube-design/components": "^1.1.8",
+    "@kube-design/components": "^1.1.11",
     "ace-builds": "^1.4.7",
     "ansi_up": "^4.0.4",
     "async-validator": "^1.8.5",
