+++ src/components/Graph/Card/index.jsx
@@ -154,12 +154,9 @@ export default class Card extends React.PureComponent {
         </div>
         <div className={styles.wrapper}>
           {this.renderDetail()}
-          <div className={styles.componentName}>{data.app}</div>
-          {data.version && (
-            <div className={styles.version}>
-              <span>version</span> {data.version}
-            </div>
-          )}
+          <div className={styles.workload}>
+            <span>{t('Deployment')}</span> {data.workload}
+          </div>
         </div>
       </div>
     )
