+++ src/components/Graph/Detail/index.jsx
@@ -85,6 +85,7 @@ export default class AppDetail extends React.Component {
           <Collapse
             activeKey={this.state.activeKey}
             onChange={this.handleCollapseChange}
+            accordion
           >
             <CollapseItem label={t('Traffic Monitoring')} key="monitor">
               <Monitors
