+++ src/components/Graph/Detail/index.scss
@@ -9,7 +9,7 @@
 }
 
 .wrapper {
-  width: 320px;
+  width: 400px;
   height: 100%;
   background-color: #fff;
   border-radius: 4px;
