+++ src/pages/projects/components/Modals/CreateApp/index.jsx
@@ -145,13 +145,17 @@ export default class ServiceDeployAppModal extends React.Component {
         'metadata.annotations["nginx.ingress.kubernetes.io/upstream-vhost"]',
         `productpage.${namespace}.svc.cluster.local`
       )
-      set(
-        formData.ingress,
-        'spec.rules[0].host',
-        gateway.isHostName
-          ? gateway.defaultIngress
-          : `productpage.${namespace}.${gateway.defaultIngress}.nip.io`
-      )
+      if (!gateway.defaultIngress) {
+        set(formData.ingress, 'spec.rules', [])
+      } else {
+        set(
+          formData.ingress,
+          'spec.rules[0].host',
+          gateway.isHostName
+            ? gateway.defaultIngress
+            : `productpage.${namespace}.${gateway.defaultIngress}.nip.io`
+        )
+      }
 
       this.setState({ formData })
     })
