+++ src/pages/projects/containers/Applications/CRDAppDetail/Tracing/index.jsx
@@ -20,14 +20,11 @@ import { isEmpty } from 'lodash'
 import React from 'react'
 import { toJS, when } from 'mobx'
 import { observer, inject } from 'mobx-react'
-import { parse } from 'qs'
 import { joinSelector } from 'utils'
 
-import { Button, Select } from '@kube-design/components'
+import { Button, Icon, Select } from '@kube-design/components'
 import TracingCard from 'projects/components/Cards/Tracing'
 import TracingDetail from 'projects/components/Modals/TracingDetail'
-
-import ServiceSelect from './ServiceSelect'
 
 import styles from './index.scss'
 
@@ -39,8 +36,6 @@ export default class Tracing extends React.Component {
 
     this.detailStore = props.detailStore
 
-    const query = parse(location.search.slice(1))
-
     this.state = {
       showDetailModal: false,
       selectItem: {},
@@ -49,7 +44,6 @@ export default class Tracing extends React.Component {
         lookback: '1h',
         limit: 5,
       },
-      defaultService: query.service,
     }
   }
 
@@ -85,6 +79,14 @@ export default class Tracing extends React.Component {
       { label: t('Last {hour} hours', { hour: 24 }), value: '24h' },
       { label: t('Last {day} days', { day: 2 }), value: '2d' },
     ]
+  }
+
+  get services() {
+    return this.detailStore.serviceStore.list.data.map(item => ({
+      label: item.name,
+      value: item.name,
+      type: item.type,
+    }))
   }
 
   getData() {
@@ -126,8 +128,8 @@ export default class Tracing extends React.Component {
     this.setState({ showDetailModal: false, selectItem: {} })
   }
 
-  handleServiceChange = service => {
-    this.setState({ serviceName: service.name }, () => {
+  handleServiceChange = value => {
+    this.setState({ serviceName: value }, () => {
       this.fetchTracing()
     })
   }
@@ -154,22 +156,23 @@ export default class Tracing extends React.Component {
     )
   }
 
-  renderServices() {
-    const { isLoading, data } = this.detailStore.serviceStore.list
-    return (
-      <ServiceSelect
-        defaultService={this.state.defaultService}
-        options={toJS(data)}
-        isLoading={isLoading}
-        onChange={this.handleServiceChange}
-      />
-    )
-  }
+  serviceRenderer = option => (
+    <span>
+      {t('Service')}: {option.label}
+    </span>
+  )
 
   renderOperations() {
     const { query } = this.state
     return (
       <div className={styles.operations}>
+        <Select
+          options={this.services}
+          value={this.state.serviceName}
+          prefixIcon={<Icon name="appcenter" />}
+          onChange={this.handleServiceChange}
+          valueRenderer={this.serviceRenderer}
+        />
         <Select
           value={query.lookback}
           options={this.lookbackOptions}
@@ -206,7 +209,6 @@ export default class Tracing extends React.Component {
   render() {
     return (
       <div>
-        {this.renderServices()}
         {this.renderTracing()}
         <TracingDetail
           detail={this.state.selectItem}
